moved filters outside of data
