+++ src/MuonDataLib/GUI/amp/presenter.py
@@ -1,7 +1,7 @@
 from MuonDataLib.GUI.table.presenter import PresenterTemplate
 from MuonDataLib.GUI.amp.view import AmplitudeView
 from MuonDataLib.GUI.plot_area.presenter import PlotAreaPresenter
-from MuonDataLib.data.filters import PeakProperty
+from MuonDataLib.filters import PeakProperty
 
 
 class AmpPresenter(PresenterTemplate):
